feat(wip): common submission manager
