+++ app/Services/QueueSubmissionManager.php
@@ -6,12 +6,9 @@ use App\Facades\Notifications;
 use App\Facades\Settings;
 use App\Models\Character\Character;
 use App\Models\Currency\Currency;
-use App\Models\Item\Item;
-use App\Models\Loot\LootTable;
 use App\Models\Queue\Queue;
 use App\Models\Queue\QueueSubmission;
 use App\Models\Queue\QueueSubmissionCharacter;
-use App\Models\Raffle\Raffle;
 use App\Models\User\User;
 use App\Models\User\UserItem;
 use Carbon\Carbon;
@@ -244,7 +241,7 @@ class QueueSubmissionManager extends Service {
             }
             $qAssets = $assets['queue'];
             // Remove queue-only rewards
-            $queueRewards = $this->removeQueueAttachments($submission);
+            $queueRewards = $this->removeSubmissionAttachments($submission);
 
             if ($user->id != $submission->user_id) {
                 // The only things we need to set are:
@@ -611,286 +608,5 @@ class QueueSubmissionManager extends Service {
         }
 
         return $this->rollbackReturn(false);
-    }
-
-    /**************************************************************************************************************
-     *
-     * ATTACHMENT FUNCTIONS
-     *
-     **************************************************************************************************************/
-
-    /**
-     * Creates user attachments for a submission.
-     *
-     * @param mixed $submission the submission object
-     * @param mixed $data       the data for creating the attachments
-     * @param mixed $user       the user object
-     */
-    private function createUserAttachments($submission, $data, $user) {
-        $userAssets = createAssetsArray();
-
-        // Attach items. Technically, the user doesn't lose ownership of the item - we're just adding an additional holding field.
-        // We're also not going to add logs as this might add unnecessary fluff to the logs and the items still belong to the user.
-        if (isset($data['stack_id'])) {
-            foreach ($data['stack_id'] as $stackId) {
-                $stack = UserItem::with('item')->find($stackId);
-                if (!$stack || $stack->user_id != $user->id) {
-                    throw new \Exception('Invalid item selected.');
-                }
-                if (!isset($data['stack_quantity'][$stackId])) {
-                    throw new \Exception('Invalid quantity selected.');
-                }
-                $stack->submission_count += $data['stack_quantity'][$stackId];
-                $stack->save();
-
-                addAsset($userAssets, $stack, $data['stack_quantity'][$stackId]);
-            }
-        }
-
-        // Attach currencies.
-        if (isset($data['currency_id'])) {
-            foreach ($data['currency_id'] as $holderKey => $currencyIds) {
-                $holder = explode('-', $holderKey);
-                $holderType = $holder[0];
-                $holderId = $holder[1];
-
-                $holder = User::find($holderId);
-
-                $currencyManager = new CurrencyManager;
-                foreach ($currencyIds as $key => $currencyId) {
-                    $currency = Currency::find($currencyId);
-                    if (!$currency) {
-                        throw new \Exception('Invalid currency selected.');
-                    }
-                    if ($data['currency_quantity'][$holderKey][$key] < 0) {
-                        throw new \Exception('Cannot attach a negative amount of currency.');
-                    }
-                    if (!$currencyManager->debitCurrency($holder, null, null, null, $currency, $data['currency_quantity'][$holderKey][$key])) {
-                        throw new \Exception('Invalid currency/quantity selected.');
-                    }
-
-                    addAsset($userAssets, $currency, $data['currency_quantity'][$holderKey][$key]);
-                }
-            }
-        }
-
-        // Get a list of rewards, then create the submission itself
-        $queueRewards = createAssetsArray();
-        if ($submission->status == 'Pending' && isset($submission->queue_id) && $submission->queue_id) {
-            if ($submission->queue->rewards) {
-                foreach ($submission->queue->rewardItems as $key => $type) {
-                    foreach ($type as $asset) {
-                        addAsset($queueRewards, $asset['asset'], $asset['quantity']);
-                    }
-                }
-            }
-        }
-        $queueRewards = mergeAssetsArrays($queueRewards, $this->processRewards($data, false));
-
-        return [
-            'userAssets'   => $userAssets,
-            'queueRewards' => $queueRewards,
-        ];
-    }
-
-    /**
-     * Creates character attachments for a submission.
-     *
-     * @param mixed $submission the submission object
-     * @param mixed $data       the data for creating character attachments
-     * @param mixed $service
-     * @param mixed $user
-     */
-    private function createCharacterAttachments($submission, $data, $service, $user) {
-        DB::beginTransaction();
-
-        try {
-            // The character identification comes in both the slug field and as character IDs
-            // that key the reward ID/quantity arrays.
-            // We'll need to match characters to the rewards for them.
-            // First, check if the characters are accessible to begin with.
-            if (isset($data['slug'])) {
-                $characters = Character::myo(0)->visible()->whereIn('slug', $data['slug'])->get();
-                if (count($characters) != count($data['slug'])) {
-                    throw new \Exception('One or more of the selected characters do not exist.');
-                }
-            } else {
-                $characters = [];
-            }
-            // process any relevant data
-            if (method_exists($service, 'processCharacters')) {
-                if (!$chardata = $service->processCharacters($submission->queue, $data, $submission)) {
-                    foreach ($service->errors()->getMessages()['error'] as $error) {
-                        flash($error)->error();
-                    }
-                    throw new \Exception('Failed to handle submission characters.');
-                }
-            }
-            // Attach characters
-            foreach ($characters as $c) {
-                // and for a specific character
-                if (method_exists($service, 'processCharacterAttachments')) {
-                    if (!$assets = $service->processCharacterAttachments($submission->queue, $data + ['character_id' => $c->id], $submission)) {
-                        foreach ($service->errors()->getMessages()['error'] as $error) {
-                            flash($error)->error();
-                        }
-                        throw new \Exception('Failed to handle submission characters.');
-                    }
-                }
-                // Now we have a clean set of assets (redundant data is gone, duplicate entries are merged)
-                // so we can attach the character to the submission
-                QueueSubmissionCharacter::create([
-                    'character_id'        => $c->id,
-                    'queue_submission_id' => $submission->id,
-                    'data'                => method_exists($service, 'finalizeCharacterAttachments') ? $service->finalizeCharacterAttachments($submission->queue, $data + ['character_id' => $c->id], $submission, $user) : null,
-                ]);
-            }
-
-            return $this->commitReturn(true);
-        } catch (\Exception $e) {
-            $this->setError('error', $e->getMessage());
-        }
-
-        return $this->rollbackReturn(false);
-    }
-
-    /**
-     * Removes attachments from a submission.
-     *
-     * @param mixed $submission the submission object
-     */
-    private function removeAttachments($submission) {
-        // This occurs when a draft is edited or rejected.
-
-        // Return all added items
-        $addonData = $submission->data['user'];
-        if (isset($addonData['user_items'])) {
-            foreach ($addonData['user_items'] as $userItemId => $quantity) {
-                $userItemRow = UserItem::find($userItemId);
-                if (!$userItemRow) {
-                    throw new \Exception('Cannot return an invalid item. ('.$userItemId.')');
-                }
-                if ($userItemRow->submission_count < $quantity) {
-                    throw new \Exception('Cannot return more items than was held. ('.$userItemId.')');
-                }
-                $userItemRow->submission_count -= $quantity;
-                $userItemRow->save();
-            }
-        }
-
-        // And currencies
-        $currencyManager = new CurrencyManager;
-        if (isset($addonData['currencies']) && $addonData['currencies']) {
-            foreach ($addonData['currencies'] as $currencyId => $quantity) {
-                $currency = Currency::find($currencyId);
-                if (!$currency) {
-                    throw new \Exception('Cannot return an invalid currency. ('.$currencyId.')');
-                }
-                if (!$currencyManager->creditCurrency(null, $submission->user, null, null, $currency, $quantity)) {
-                    throw new \Exception('Could not return currency to user. ('.$currencyId.')');
-                }
-            }
-        }
-    }
-
-    /**
-     * Processes reward data into a format that can be used for distribution.
-     *
-     * @param array $data
-     * @param bool  $isCharacter
-     * @param bool  $isStaff
-     *
-     * @return array
-     */
-    private function processRewards($data, $isCharacter, $isStaff = false) {
-        if ($isCharacter) {
-            $assets = createAssetsArray(true);
-
-            if (isset($data['character_currency_id'][$data['character_id']]) && isset($data['character_quantity'][$data['character_id']])) {
-                foreach ($data['character_currency_id'][$data['character_id']] as $key => $currency) {
-                    if ($data['character_quantity'][$data['character_id']][$key]) {
-                        addAsset($assets, $data['currencies'][$currency], $data['character_quantity'][$data['character_id']][$key]);
-                    }
-                }
-            } elseif (isset($data['character_rewardable_type'][$data['character_id']]) && isset($data['character_rewardable_id'][$data['character_id']]) && isset($data['character_rewardable_quantity'][$data['character_id']])) {
-                $data['character_rewardable_id'] = array_map([$this, 'innerNull'], $data['character_rewardable_id']);
-
-                foreach ($data['character_rewardable_id'][$data['character_id']] as $key => $reward) {
-                    switch ($data['character_rewardable_type'][$data['character_id']][$key]) {
-                        case 'Currency':if ($data['character_rewardable_quantity'][$data['character_id']][$key]) {
-                            addAsset($assets, $data['currencies'][$reward], $data['character_rewardable_quantity'][$data['character_id']][$key]);
-                        }break;
-                        case 'Item':if ($data['character_rewardable_quantity'][$data['character_id']][$key]) {
-                            addAsset($assets, $data['items'][$reward], $data['character_rewardable_quantity'][$data['character_id']][$key]);
-                        }break;
-                        case 'LootTable':if ($data['character_rewardable_quantity'][$data['character_id']][$key]) {
-                            addAsset($assets, $data['tables'][$reward], $data['character_rewardable_quantity'][$data['character_id']][$key]);
-                        }break;
-                    }
-                }
-            }
-
-            return $assets;
-        } else {
-            $assets = createAssetsArray(false);
-            // Process the additional rewards
-            if (isset($data['rewardable_type']) && $data['rewardable_type']) {
-                foreach ($data['rewardable_type'] as $key => $type) {
-                    $reward = null;
-                    switch ($type) {
-                        case 'Item':
-                            $reward = Item::find($data['rewardable_id'][$key]);
-                            break;
-                        case 'Currency':
-                            $reward = Currency::find($data['rewardable_id'][$key]);
-                            if (!$reward->is_user_owned) {
-                                throw new \Exception('Invalid currency selected.');
-                            }
-                            break;
-                        case 'LootTable':
-                            if (!$isStaff) {
-                                break;
-                            }
-                            $reward = LootTable::find($data['rewardable_id'][$key]);
-                            break;
-                        case 'Raffle':
-                            if (!$isStaff) {
-                                break;
-                            }
-                            $reward = Raffle::find($data['rewardable_id'][$key]);
-                            break;
-                    }
-                    if (!$reward) {
-                        continue;
-                    }
-                    addAsset($assets, $reward, $data['quantity'][$key]);
-                }
-            }
-
-            return $assets;
-        }
-    }
-
-    /**
-     * Removes the attachments associated with a queue from a submission.
-     *
-     * @param mixed $submission the submission object
-     */
-    private function removeQueueAttachments($submission) {
-        $assets = $submission->data;
-        // Get a list of rewards, then create the submission itself
-        $queueRewards = createAssetsArray();
-        $queueRewards = mergeAssetsArrays($queueRewards, parseAssetData($assets['rewards']));
-        if (isset($submission->queue_id) && $submission->queue_id) {
-            if ($submission->queue->rewards) {
-                foreach ($submission->queue->rewardItems as $key => $type) {
-                    foreach ($type as $asset) {
-                        removeAsset($queueRewards, $asset['asset'], $asset['quantity']);
-                    }
-                }
-            }
-        }
-
-        return $queueRewards;
     }
 }
